make functions compatible with very large values

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
  */
 int divide(unsigned long int num, unsigned int div)
 {
-	unsigned int dividend = 0;
+	unsigned long int dividend = 0;
 
 	while (num >= div)
 	{
@@ -30,7 +30,7 @@ int divide(unsigned long int num, unsigned int div)
  */
 unsigned int find_no_digit(unsigned long int num)
 {
-	unsigned int count = 0;
+	unsigned long int count = 0;
 
 	while (num >= 2)
 	{
