+++ 0x14-bit_manipulation/3-set_bit.c
@@ -9,7 +9,7 @@
  */
 int divide(unsigned long int num, unsigned int div)
 {
-	unsigned int dividend = 0;
+	unsigned long int dividend = 0;
 
 	while (num >= div)
 	{
