+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,7 +12,7 @@
  */
 int divide(unsigned long int num, unsigned int div)
 {
-	int dividend = 0;
+	unsigned long int dividend = 0;
 
 	while (num >= div)
 	{
